requirements.txt
torch
torchvision
Pillow
faiss-cpu
numpy
scikit-learn
tqdm
